WIP: Positive error description
Currently AllowValueMatcher has a commented-out example usage, which should be
uncommented when this is ready.

Ideally the `if expected_message` nil-check in AllowValueMatcher will be handled
by PositiveErrorDescription too.

Use it in FormattedErrorDescription as well.

We use PositiveErrorDescription#matches? and #matched_error to grab the
description and return the correct thing in FormattedErrorDescription. I
envision FormattedErrorDescription being used in AllowValueMatcher too, probably
with a #matches? method too that delegates.